ブラックリスト指定されたAPIか否かを判定する処理を、Blacklistクラスに移動
#3
